Split printer actions from printer status card
