+++ makerion_web/lib/makerion_web/views/printer_action_view.ex
@@ -1,0 +1,6 @@
+defmodule MakerionWeb.PrinterActionView do
+  use MakerionWeb, :view
+
+  def button_state(false), do: "disabled"
+  def button_state(true), do: ""
+end
