+++ makerion_web/lib/makerion_web/views/printer_status_view.ex
@@ -1,6 +1,3 @@
 defmodule MakerionWeb.PrinterStatusView do
   use MakerionWeb, :view
-
-  def button_state(false), do: "disabled"
-  def button_state(true), do: ""
 end
